+++ examples/example_3.py
@@ -1,12 +1,12 @@
 '''
 Example 3: Reading temperature from a sensor with retry and custom handling
 
-In this example, we will use the with_retry() decorator to implement a function that reads 
+In this example, we will use the resilient_call() decorator to implement a function that reads 
 the temperature from a sensor and retries the read operation in case of failure. We will also use 
 all the available options to customize the handling of different temperature values.
 '''
 import random
-from resilenter_caller import with_retry, RETRY_EVENT
+from resilenter_caller import resilient_call, RETRY_EVENT
 
 class SensorError(Exception):
     pass
@@ -30,7 +30,7 @@ def handle_valid_temperature(temp):
     print(f"Temperature is within the valid range: {temp}")
     return temp
 
-@with_retry()
+@resilient_call()
 def get_temperature_data():
     return read_temperature_from_sensor()
 
@@ -39,7 +39,7 @@ if __name__ == "__main__":
         retries=5,
         delay=2,
         exceptions={SensorError: RETRY_EVENT},
-        opts={-1: handle_low_temperature, 1: handle_high_temperature},
-        opts_criteria=lambda temp: -1 if temp < 0 else 1 if temp > 30 else 0,
+        conditions={-1: handle_low_temperature, 1: handle_high_temperature},
+        conditions_criteria=lambda temp: -1 if temp < 0 else 1 if temp > 30 else 0,
         on_retry=lambda tries: print(f"Retry {tries}")
     )