+++ examples/example_7.py
@@ -2,7 +2,7 @@
 Example 7: Scrape all the articles from glizzykingdreko's medium account
 
 In this example, we will create a SimpleScraper class that monitors glizzykingdreko's 
-Medium blog for new articles. By defining the start method with the @with_retry() 
+Medium blog for new articles. By defining the start method with the @resilient_call() 
 decorator and passing delay=5 and exceptions={"all": RETRY_EVENT} when calling it, 
 we ensure that the scraper handles all exceptions with a retry and adds a delay of 
 5 seconds between each request.
@@ -11,7 +11,7 @@ from bs4 import BeautifulSoup
 from requests import Session
 from typing import List
 from time import sleep
-from resilenter_caller import with_retry, RETRY_EVENT
+from resilenter_caller import resilient_call, RETRY_EVENT
 
 class SimpleScraper:
     def __init__(self):
@@ -20,7 +20,7 @@ class SimpleScraper:
         # and handling any exception with a retry event.
         self.start(delay=5, exceptions={"all": RETRY_EVENT})
 
-    @with_retry()
+    @resilient_call()
     def start(self) -> None:
         data = self.load_api_data()
         self.load_response_details(data)